Export va Importing comandalari qo'shildi

+++ utils.py
@@ -1,7 +1,9 @@
 import ast
 import hashlib
+import json
 import os
 import shutil
+import sqlite3
 import traceback
 from pathlib import Path
 from sqlite3.dbapi2 import Timestamp
@@ -56,6 +58,7 @@ class Utils:
         self, db: DataBase, message: str = "", path: str = "."
     ) -> Tuple[str, bool]:
         path = self.clear_path(path)
+        print(path)
         if not db.project_exists(path):
             return "Project not found!", False
 
@@ -63,7 +66,7 @@ class Utils:
         project_flow: str
         parent_id: int
 
-        project_id, project_flow, parent_id = db.get_project_id(path)
+        project_id, project_flow, parent_id, *_data = db.get_project_id(path)
 
         uid = str(uuid4())
         db.create_tree(uid)
@@ -73,6 +76,7 @@ class Utils:
             "mid": set(),
             "right": set(),
             "full": set(),
+            "folders": {".chillbek", ".venv"},
             "changed": False,
         }
 
@@ -89,14 +93,21 @@ class Utils:
                 if not is_success:
                     return self.chillignore["error"], False
 
+            if self.is_folder_ignored(
+                set(Path(root.replace(path, "", count=1)).parts), self.chillignore
+            ):
+                continue
+
+            tree_root = Path(root).relative_to(Path(path)).as_posix()
+            saved_files = []
             for file in files:
                 if self.chillignore["changed"] and self.is_file_ignored(
                     file, self.chillignore
                 ):
                     continue
 
-                path = os.path.join(root, file)
-                hash: str = self.hashing_file(path)
+                file_path = os.path.join(root, file)
+                hash: str = self.hashing_file(file_path)
                 hash_path: str = os.path.join(hash[:3], hash[3:])
 
                 if not db.blob_exists(hash):
@@ -108,24 +119,31 @@ class Utils:
                     try:
                         if not os.path.exists(os.path.join(dst, hash_path)):
                             shutil.copy2(
-                                path,
+                                file_path,
                                 os.path.join(dst, hash_path),
                                 follow_symlinks=False,
                             )
-                    except shutil.SameFileError:
+                    except Exception as e:
+                        print(f"Error in moving file {path} :", str(e))
                         pass
 
                 item_id = db.get_blob_id(hash_path)
+                saved_files.append(file)
+                tree[0].append([tree_root, file, item_id])
 
-                tree[0].append([root, file, item_id])
-
-            tree[1].append([root, dirs, files])
+            _dirs: set = self.chillignore["folders"].copy()
+            dirs = [dir for dir in dirs if len(dirs) == len(_dirs | set(dir))]
+            tree[1].append([tree_root, dirs, saved_files])
 
         tree_hash = hashlib.sha256(str(str(tree) + project_flow).encode()).hexdigest()
         if db.tree_exists(tree_hash):
             db.delete_tree(tree_id)
-            return "This tree already have!", False
-        db.update_tree(tree_hash, str(tree), tree_id)
+            tree_id = db.get_tree_id(tree_hash)
+            if self.search_tree(db, project_flow, project_id):
+                return "This tree already has in this flow", False
+
+        else:
+            db.update_tree(tree_hash, str(tree), tree_id)
 
         if message == "":
             message = f"Version 0.{len(saves) if len(saves) > 0 else 1}v and flow {project_flow}"
@@ -152,21 +170,23 @@ class Utils:
             chill_base_path: str = os.path.join(
                 os.path.expanduser("~"), ".ChillManager", "objects"
             )
-
-            shutil.rmtree(base_path)
+            if os.path.exists(base_path):
+                shutil.rmtree(base_path)
             for root, file, blob_id in tree_dict[0]:
-                os.makedirs(root, exist_ok=True)
+                root = os.path.join(base_path, root)
+                os.makedirs(os.path.join(base_path, root), exist_ok=True)
                 blob: str = db.get_blob_by_id(int(blob_id))
                 from_path: str = os.path.join(chill_base_path, blob)
                 to_path: str = os.path.join(root, file)
-                if not root.startswith(base_path):
+
+                if not os.path.exists(root):
                     print(f"Unexpected path found [{root}]")
                     continue
 
                 shutil.copy2(from_path, to_path)
 
             for root, *other in tree_dict[1]:
-                os.makedirs(root, exist_ok=True)
+                os.makedirs(os.path.join(base_path, root), exist_ok=True)
 
             db.update_project_head(project[0], save_id)
 
@@ -188,7 +208,7 @@ class Utils:
         project_flow: str
         save_head: int
 
-        project_id, project_flow, save_head = db.get_project_id(path)
+        project_id, project_flow, save_head, *_data = db.get_project_id(path)
 
         response: list = [
             f"Current flow [ {self.colored_print(project_flow, self.color_data['blue'])} ] and current save id [ {self.colored_print(str(save_head), self.color_data['blue'])} ]"
@@ -211,28 +231,10 @@ class Utils:
             project_data: tuple = db.get_project_by_id(
                 project_id
             )  # id path flow head_save flow_data
-            save = db.get_save_by_id(
-                project_data[3]
-            )  # id, project_id, tree_id, parent_id, comment, time, flow
-
             if flow_name == project_data[2]:
                 return "Current flow and this flow are same!", True
 
-            flow_data: dict = ast.literal_eval(project_data[-1])
-            old_tree: tuple = db.get_tree_by_id(save[2])
-
-            tree_hash: str = hashlib.sha256(
-                str(str(old_tree[-1]) + flow_name).encode()
-            ).hexdigest()
-
-            db.create_tree(tree_hash)
-            tree_id = db.get_tree_id(tree_hash)
-            db.update_tree(tree_hash, old_tree[-1], tree_id)
-
-            flow_data[project_data[2]] = int(project_data[3])
-            db.update_project_flow_data(project_id, flow_data)
-
-            db.update_project_head(project_id, flow_data[flow_name])
+            db.update_project_head(project_id, 0)
             db.update_project_flow(project_id, flow_name)
 
             return "Flow changed!", True
@@ -248,9 +250,7 @@ class Utils:
             )  # id path flow head_save flow_data
 
             flow_data: dict = ast.literal_eval(project_data[-1])
-
-            new_save_id: int = self.copy_save(db, project_data[3], flow_name)
-            flow_data[flow_name] = new_save_id
+            flow_data[flow_name] = 0
             db.update_project_flow_data(project_id, str(flow_data))
 
             return "Flow created!", True
@@ -258,6 +258,230 @@ class Utils:
         except Exception as e:
             traceback.print_exc()
             return e, False
+
+    def clear_base(self, db: DataBase):
+        try:
+            data = {
+                "projects": set(),
+                "trees": set(),
+                "blobs": set(),
+            }
+
+            projects: list = db.get_list_projects()
+            for row in projects:
+                data["projects"].add(int(row[0]))
+
+            projects_len: int = len(data["projects"])
+            saves: list = db.get_list_saves()
+            for row in saves:
+                pr_id = row[1]
+                pr_data: set = data["projects"].copy()
+                pr_data.add(pr_id)
+                if len(pr_data) > projects_len:
+                    db.delete_save(row[0])
+                    print("Save: ", row[0])
+                    projects_len += 1
+                    continue
+                data["trees"].add(row[2])
+
+            trees: list = db.get_list_trees()
+            in_use_trees: set = data["trees"]
+            all_trees: set = set([row[0] for row in trees])
+            deleting_tree: set = all_trees - in_use_trees
+
+            for id in deleting_tree:
+                db.delete_tree(id)
+                print("Tree: ", id)
+
+            for row in trees:
+                raw_tree: str = row[2]
+                rb_tree: list = ast.literal_eval(raw_tree)
+
+                blobs: list = rb_tree[0]
+                for blob in blobs:
+                    data["blobs"].add(blob[2])
+
+            in_use_blobs: set = data["blobs"]
+            all_blobs: set = set([row[0] for row in db.get_list_blobs()])
+            deleting_blobs: set = all_blobs - in_use_blobs
+            for id in deleting_blobs:
+                db.delete_blob(id)
+                print("Blob: ", id)
+
+            return "Success", True
+
+        except Exception as e:
+            return e, False
+
+    def export_project(self, db: DataBase, project_id: int) -> tuple:
+        try:
+            # id, path, flow, head_save, flow_data
+            project_data: tuple = db.get_project_by_id(project_id)
+            chill_base_path: str = os.path.join(
+                os.path.expanduser("~"), ".ChillManager", "objects"
+            )
+            chillbek_path: str = os.path.join(project_data[1], ".chillbek")
+            blob_files_path: str = os.path.join(chillbek_path, "blobs")
+
+            if os.path.exists(chillbek_path):
+                shutil.rmtree(chillbek_path)
+            os.makedirs(blob_files_path)
+
+            data: dict = {"project": project_data}
+
+            saves: list = db.get_list_saves(project_id)
+            blobs: set = set()
+            trees: list = []
+
+            for save in saves:
+                tree_data: tuple = db.get_tree_by_id(save[2])
+                tree: list = ast.literal_eval(tree_data[2])
+                for blob_data in tree[0]:
+                    blob: str = db.get_blob_by_id(int(blob_data[-1]))
+                    blobs.add((int(blob_data[-1]), blob))
+
+                    from_path: str = os.path.join(chill_base_path, blob)
+                    to_path: str = os.path.join(blob_files_path, blob)
+                    os.makedirs(os.path.join(blob_files_path, blob[:3]), exist_ok=True)
+                    try:
+                        shutil.copy2(from_path, to_path)
+                    except Exception as e:
+                        print("Problem with this file!", from_path)
+                        continue
+
+                for file_data in tree[1]:
+                    file_data[0] = str(file_data[0]).replace(
+                        str(project_data[1]), ".", 1
+                    )
+
+                trees.append((tree_data[0], tree_data[1], tree))
+
+            data["saves"] = saves
+            data["trees"] = trees
+            data["blobs"] = list(blobs)
+
+            with open(os.path.join(chillbek_path, "data.json"), "w") as outfile:
+                json.dump(data, outfile)
+
+            return "Project ready for export!", True
+
+        except Exception as e:
+            traceback.print_exc()
+            return e, False
+
+    def import_project(self, db: DataBase, path: str) -> tuple:
+        try:
+            path = self.clear_path(path)
+            chillbek_path = os.path.join(path, ".chillbek")
+            data_file = os.path.join(chillbek_path, "data.json")
+
+            if not os.path.exists(data_file):
+                return "No export data found in .chillbek folder!", False
+
+            with open(data_file, "r") as f:
+                data = json.load(f)
+
+            if db.project_exists(path):
+                inp = input(
+                    "Project already exists in database!\nChoose (Y)es for delete and importing or Other for stop protcess!"
+                )
+                if inp.lower() == "yes" or inp.lower() == "y":
+                    db.delete_project(db.get_project_id(path)[0])
+                    self.clear_base(db)
+                else:
+                    return "Protcess stopped!", False
+
+            chill_base_path = os.path.join(
+                os.path.expanduser("~"), ".ChillManager", "objects"
+            )
+            blob_files_path = os.path.join(chillbek_path, "blobs")
+
+            blob_id_map = {}
+            for blob in data["blobs"]:
+                blob_hash_path = blob[1]
+                from_path = os.path.join(blob_files_path, blob_hash_path)
+                to_path = os.path.join(chill_base_path, blob_hash_path)
+                if os.path.exists(from_path):
+                    os.makedirs(os.path.dirname(to_path), exist_ok=True)
+                    try:
+                        shutil.copy2(from_path, to_path)
+                    except Exception as e:
+                        print("Error: ", str(e))
+                        continue
+                if not db.blob_exists(blob_hash_path):
+                    db.create_blob(blob_hash_path)
+                new_id = db.get_blob_id(blob_hash_path)
+                blob_id_map[blob[0]] = new_id
+
+            p_info = data["project"]
+            db.create_project(
+                path, flow=p_info[2], flow_data=ast.literal_eval(p_info[4])
+            )
+            project_id = db.get_project_id(path)[0]
+
+            tree_id_map = {}
+            for tree_item in data["trees"]:
+                content = tree_item[2]
+                for item in content[0]:
+                    try:
+                        item[-1] = blob_id_map[item[-1]]
+                    except KeyError:
+                        print("This item blob not found:", item)
+                        continue
+
+                uid = str(uuid4())
+                db.create_tree(uid)
+                new_tree_id = db.get_tree_id(uid)
+                db.update_tree(uid, str(tree_item[2]), new_tree_id)
+                tree_id_map[tree_item[0]] = new_tree_id
+
+            save_id_map = {}
+            for s in data["saves"]:
+                # s: id, project_id, tree_id, parent_id, comment, time, flow
+                old_id = s[0]
+                new_tree_id = tree_id_map.get(s[2], 0)
+                tree: tuple = db.get_tree_by_id(new_tree_id)
+                new_tree_hash: str = hashlib.sha256(
+                    str(str(tree[2]) + s[6]).encode()
+                ).hexdigest()
+                try:
+                    db.update_tree(new_tree_hash, tree[2], tree[0])
+                except sqlite3.IntegrityError:
+                    if new_tree_id != -1:
+                        db.delete_tree(new_tree_id)
+                        new_tree_id = db.get_tree_id(new_tree_hash)
+
+                db.create_save(project_id, new_tree_id, 0, s[4], s[5], s[6])
+
+                new_save_id = db.get_save_id(
+                    project_id, new_tree_id, 0, s[4], s[5], s[6]
+                )[0]
+                save_id_map[old_id] = new_save_id
+
+            for s in data["saves"]:
+                old_id = s[0]
+                old_parent_id = s[3]
+                if old_parent_id and old_parent_id != -1:
+                    new_parent_id = save_id_map.get(old_parent_id, 0)
+                    db.update_save_parent(save_id_map[old_id], new_parent_id)
+
+            new_head = save_id_map.get(p_info[3], 0)
+            db.update_project_head(project_id, new_head)
+
+            old_flow_data = ast.literal_eval(p_info[4])
+            new_flow_data = {
+                flow: save_id_map.get(sid, -1) for flow, sid in old_flow_data.items()
+            }
+            db.update_project_flow_data(project_id, str(new_flow_data))
+
+            return "Project successfully imported!", True
+
+        except Exception as e:
+            traceback.print_exc()
+            return str(e), False
+
+    def is_folder_ignored(self, folders, data):
+        return True if len(set(data["folders"]) & set(folders)) > 0 else False
 
     def is_file_ignored(self, file, data):
         def sliding_window(text, ln):
@@ -288,7 +512,13 @@ class Utils:
 
     def get_chillignore(self, path: str):
         try:
-            res: dict = {"left": set(), "mid": set(), "right": set(), "full": set()}
+            res: dict = {
+                "left": set(),
+                "mid": set(),
+                "right": set(),
+                "full": set(),
+                "folders": {".chillbek", ".venv"},
+            }
 
             with open(path, "r") as file:
                 lst: list[str] = file.read().splitlines()
@@ -296,7 +526,9 @@ class Utils:
             for line in lst:
                 if line == "*":
                     continue
-                if line.startswith("*") and line.endswith("*"):
+                if line[-1] == "/":
+                    res["folders"].add(line[:-1])
+                elif line.startswith("*") and line.endswith("*"):
                     res["mid"].add(line[1:-1])
                 elif line.startswith("*"):
                     res["right"].add(line[1:])
@@ -339,6 +571,15 @@ class Utils:
         )
         db.create_save(*save_data)
         return db.get_save_id(*save_data)[0]
+
+    def search_tree(self, db: DataBase, project_flow, project_id):
+        saves = db.get_list_saves_by_project(int(project_id))
+
+        for save in saves:
+            if save[-1] == project_flow:
+                return True
+
+        return False
 
     def colored_print(self, data: str, color: str) -> str:
         return color + str(data) + self.color_data["black"]
